# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BIKO API",
        "description": "API para gerenciamento de usuários, categorias e publicações",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://biko-api.test/api",
            "description": "Servidor da API"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/collections": {
            "get": {
                "tags": [
                    "Coleções"
                ],
                "summary": "Listar coleções do usuário autenticado",
                "operationId": "d4df76267f6c624eb9f318b4e6e0f312",
                "responses": {
                    "200": {
                        "description": "Lista de coleções",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Collection"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Coleções"
                ],
                "summary": "Criar coleção",
                "operationId": "e14763de4129762a96fd54451189b049",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Eletricistas"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Coleção criada com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/collections/{collection}": {
            "get": {
                "tags": [
                    "Coleções"
                ],
                "summary": "Exibir coleção com publicações",
                "operationId": "53eb5876f82ae6d63b50ede5620f99e6",
                "parameters": [
                    {
                        "name": "collection",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coleção encontrada"
                    },
                    "404": {
                        "description": "Coleção não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Coleções"
                ],
                "summary": "Renomear coleção",
                "operationId": "9d51d8672715bd32c0232c1455fcbb37",
                "parameters": [
                    {
                        "name": "collection",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Favoritos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Coleção atualizada com sucesso"
                    },
                    "403": {
                        "description": "Não autorizado (coleção padrão não pode ser renomeada)"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Coleções"
                ],
                "summary": "Deletar coleção",
                "operationId": "466d29c63d3de005fe2ea36d32bd6b3b",
                "parameters": [
                    {
                        "name": "collection",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coleção deletada com sucesso"
                    },
                    "403": {
                        "description": "Não autorizado (coleção padrão não pode ser deletada)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/collections/{collection}/publications/{publication}": {
            "post": {
                "tags": [
                    "Coleções"
                ],
                "summary": "Salvar/remover publicação da coleção (toggle)",
                "operationId": "b564631f0785893109ee2764315ef871",
                "parameters": [
                    {
                        "name": "collection",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "publication",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Publicação salva/removida da coleção"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Eletricista"
                    },
                    "slug": {
                        "type": "string",
                        "example": "eletricista"
                    }
                },
                "type": "object"
            },
            "UserImage": {
                "title": "UserImage",
                "properties": {
                    "url": {
                        "type": "string",
                        "example": "http://localhost/storage/1/avatar.jpg"
                    },
                    "thumb_url": {
                        "type": "string",
                        "example": "http://localhost/storage/1/conversions/avatar-thumb.jpg",
                        "nullable": true
                    }
                },
                "type": "object",
                "nullable": true
            },
            "User": {
                "title": "User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Hayssa Gomes"
                    },
                    "email": {
                        "type": "string",
                        "example": "hayssa@email.com"
                    },
                    "avatar": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserImage"
                            }
                        ],
                        "nullable": true
                    },
                    "cover": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserImage"
                            }
                        ],
                        "nullable": true
                    },
                    "followers_count": {
                        "type": "integer",
                        "example": 12
                    },
                    "following_count": {
                        "type": "integer",
                        "example": 8
                    }
                },
                "type": "object"
            },
            "Publication": {
                "title": "Publication",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "Preciso de um eletricista para instalação residencial"
                    },
                    "type": {
                        "description": "0=cliente, 1=prestador",
                        "type": "integer",
                        "example": 0
                    },
                    "author": {
                        "$ref": "#/components/schemas/User"
                    },
                    "city": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "São Paulo"
                            }
                        },
                        "type": "object"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "urgente",
                            "residencial"
                        ]
                    },
                    "media": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PublicationMedia"
                        }
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comment"
                        }
                    },
                    "likes_count": {
                        "type": "integer",
                        "example": 5
                    },
                    "comments_count": {
                        "type": "integer",
                        "example": 3
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Comment": {
                "title": "Comment",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "comment": {
                        "type": "string",
                        "example": "Ótimo serviço!"
                    },
                    "author": {
                        "$ref": "#/components/schemas/User"
                    },
                    "media": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PublicationMedia"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "PublicationMedia": {
                "title": "PublicationMedia",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "foto-servico.jpg"
                    },
                    "mime_type": {
                        "type": "string",
                        "example": "image/jpeg"
                    },
                    "size": {
                        "description": "Tamanho em bytes",
                        "type": "integer",
                        "example": 204800
                    },
                    "url": {
                        "type": "string",
                        "example": "http://localhost/storage/1/foto-servico.jpg"
                    },
                    "thumb_url": {
                        "description": "URL do thumbnail (null para vídeos)",
                        "type": "string",
                        "example": "http://localhost/storage/1/conversions/foto-servico-thumb.jpg",
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "image",
                            "video"
                        ],
                        "example": "image"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Collection": {
                "title": "Collection",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Salvos"
                    },
                    "is_default": {
                        "type": "boolean",
                        "example": true
                    },
                    "publications_count": {
                        "type": "integer",
                        "example": 5
                    },
                    "publications": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Publication"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Informe o token no formato: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}